+++ source/extensions/isaacsim.sensors.physx.ui/isaacsim/sensors/physx/ui/menu.py
@@ -26,20 +26,6 @@ class RangeSensorMenu:
                     make_menu_item_description(ext_id, "Generic", lambda a=weakref.proxy(self): a._add_generic()),
                 ],
             ),
-            MenuItemDescription(
-                name="Ultrasonic",
-                sub_menu=[
-                    make_menu_item_description(
-                        ext_id, "Array", lambda a=weakref.proxy(self): a._add_ultrasonic_array()
-                    ),
-                    make_menu_item_description(
-                        ext_id, "Emitter", lambda a=weakref.proxy(self): a._add_ultrasonic_emitter()
-                    ),
-                    make_menu_item_description(
-                        ext_id, "FiringGroup", lambda a=weakref.proxy(self): a._add_ultrasonic_firing_group()
-                    ),
-                ],
-            ),
         ]
 
         self._menu_items = [
@@ -76,43 +62,6 @@ class RangeSensorMenu:
             high_lod=False,
             yaw_offset=0.0,
             enable_semantics=False,
-        )
-
-    def _add_ultrasonic_array(self, *args, **kwargs):
-        result, prim = omni.kit.commands.execute(
-            "RangeSensorCreateUltrasonicArray",
-            path="/UltrasonicArray",
-            parent=self._get_stage_and_path(),
-            min_range=0.4,
-            max_range=3.0,
-            draw_points=False,
-            draw_lines=False,
-            horizontal_fov=15.0,
-            vertical_fov=10.0,
-            horizontal_resolution=0.5,
-            vertical_resolution=0.5,
-            num_bins=224,
-            emitter_prims=[],
-            firing_group_prims=[],
-        )
-
-    def _add_ultrasonic_emitter(self, *args, **kwargs):
-        result, prim = omni.kit.commands.execute(
-            "RangeSensorCreateUltrasonicEmitter",
-            path="/UltrasonicEmitter",
-            parent=self._get_stage_and_path(),
-            per_ray_intensity=1.0,
-            yaw_offset=0.0,
-            adjacency_list=[],
-        )
-
-    def _add_ultrasonic_firing_group(self, *args, **kwargs):
-        result, prim = omni.kit.commands.execute(
-            "RangeSensorCreateUltrasonicFiringGroup",
-            path="/UltrasonicFiringGroup",
-            parent=self._get_stage_and_path(),
-            emitter_modes=[],
-            receiver_modes=[],
         )
 
     def _add_generic(self, *args, **kwargs):