+++ source/extensions/isaacsim.sensors.physx/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
+## [2.0.0] - 2024-10-04
+### Removed
+- Ultrasonic sensor.
+
 ## [1.0.0] - 2024-09-23
 
 ### Added
-- Merges omni.isaac.proximity_sensor, isaacsim.sensors.physx, and APIs from omni.isaac.sensor into a single sensor.
+- Merges omni.isaac.proximity_sensor, isaacsim.sensors.physx, and APIs from omni.isaac.sensor into a single extension.